examples/agentic/data/physio_signal_prep/requirements.txt
mne==1.8.0
neurokit2==0.2.7
numpy==1.26.4
pandas==2.2.3
python-dotenv==1.0.1
PyYAML==6.0.2
openai==1.51.0
requests==2.32.3
tqdm==4.66.5
